Score coverage by their abnormality

coverage_score will score disruptions in the coverage across the intronic/exonic regions associated with each junction. This abnormality score generated by score_func operates by calculating the deviation of the coverage in patients to a coverage across the same regions in controls. Then, for each junction it obtains the score of the region with the greatest disruption.

coverage_score(junctions, coverage, score_func = .zscore, ...)

Arguments

junctions
junction data as a RangedSummarizedExperiment-class object.

coverage
list containing normalised coverage data that is outputted from coverage_norm.

score_func
function to score junctions by their abnormality. By default, will use a z-score but can be switched to a user-defined function. This function must take as input an x and y argument, containing case and control counts respectively. This must return a numeric vector equal to the length of x with elements corresponding to a abnormality of each junction.

...
additional arguments passed to score_func.

Value

junctions as SummarizedExperiment object with additional assays named "coverage_region" and "coverage_score". "coverage_region" labels the region of greatest disruption (1 = exon_start, 2 = exon_end, 3 = intron) and "coverage_score" contains the abnormality scores of the region with the greatest disruption.

See also

Examples

if (FALSE) {
# leave this as not run for now to save time for R CMD check
ref <-
    "ftp://ftp.ensembl.org/pub/release-100/gtf/homo_sapiens/Homo_sapiens.GRCh38.100.gtf.gz"
ref <-
    GenomicFeatures::makeTxDbFromGFF(ref)
junctions <-
    junction_norm(junctions_annot_example)
junctions <-
    junction_score(junctions)
coverage_paths_case <-
    list.files("/data/RNA_seq_diag/mito/bw/", full.names = T)[1:2]
coverage_paths_control <-
    list.files("/data/recount/GTEx_SRP012682/gtex_bigWigs/all_gtex_tissues_raw_bigWigs/",
        full.names = T
    )[1:2]
coverage <- coverage_norm(
    junctions,
    ref,
    unannot_width = 20,
    coverage_paths_case,
    coverage_paths_control,
    coverage_chr_control = "chr"
)

junctions
}