Load junctions from patient and control RNA-seq data

junc_load loads in raw patient and control junction data and formats it into a SummarizedExperiment object. Control samples can be user-inputted or selected from GTEx data publicly released through the recount2 project (https://jhubiostatistics.shinyapps.io/recount/) and downloaded through snaptron (http://snaptron.cs.jhu.edu/). By default, junc_load expects the junction data to be in STAR aligned format (SJ.out).

junc_load(
  junc_paths,
  metadata = NULL,
  controls = rep(FALSE, length(junc_paths)),
  load_func = .load_STAR,
  chrs = NULL
)

Arguments

junc_paths
file path(s) to junction data.

metadata
dataframe containing sample metadata with rows in the same order and corresponding to file path(s). Will be used as the colData of the SummarizedExperiment object.

controls
either a logical vector of the same length as paths with TRUEs labelling controls Or, "fibroblasts" representing the samples of which GTEx tissue to use as controls. If left unchanged, by default will assume all samples are patients.

load_func
function to load in junctions. By default, requires STAR formatted junctions (SJ.out). But this can be switched dependent on the format of the user's junction data. Function must take as input a junction path then return a dataframe with the columns "chr", "start", "end", "strand" and "count".

chrs
chrs chromosomes to keep. If NULL, no filter is applied.

Value

SummarizedExperiment object containing junction data.

Examples

if (FALSE) {
example_juncs_1_path <-
    system.file("extdata", "example_juncs_1.txt",
        "dasper",
        mustWork = TRUE
    )
juncs <-
    junc_load(
        junc_paths = c(example_juncs_1_path),
        metadata = dplyr::tibble(samp_id = c("example_1"))
    )
juncs
}