Annotate junctions using reference annotation

junction_annot annotates junctions by 1. whether their start and/or end position precisely overlaps with an annotated exon boundary and 2. whether that junction matches an intron definition from existing annotation. Using this information along with the strand, junctions are categorised into "annotated", "novel_acceptor", "novel_donor", "novel_combo", "novel_exon_skip", "ambig_gene" and "unannotated".

junction_annot(junctions, ref)

Arguments

junctions
junction data as a RangedSummarizedExperiment-class object.

ref
either path to gtf/gff3 or object of class TxDb-class.

Value

junctions as a RangedSummarizedExperiment-class object with additional rowData detailing overlapping genes/transcripts/exons and junction categories.

See also

Examples

if (!exists("ref")) {
    ref <- "ftp://ftp.ensembl.org/pub/release-100/gtf/homo_sapiens/Homo_sapiens.GRCh38.100.gtf.gz"
    ref <- GenomicFeatures::makeTxDbFromGFF(ref)
}

#> Import genomic features from the file as a GRanges object ... 

#> OK

#> Prepare the 'metadata' data frame ... 

#> OK

#> Make the TxDb object ... 

#> Warning: The "phase" metadata column contains non-NA values for features of type
#>   stop_codon. This information was ignored.

#> OK

if (!exists("junctions_annoted")) {
    junctions_annoted <-
        junction_annot(
            junctions_example,
            ref
        )
}

#> [1] "2020-10-03 14:08:31 - Obtaining co-ordinates of annotated exons and junctions from gtf/gff3..."
#> [1] "2020-10-03 14:08:44 - Getting junction annotation using overlapping exons..."
#> [1] "2020-10-03 14:08:46 - Tidying junction annotation..."
#> [1] "2020-10-03 14:08:46 - Deriving junction categories..."
#> [1] "2020-10-03 14:08:52 - done!"

junctions_annoted

#> class: RangedSummarizedExperiment 
#> dim: 97671 5 
#> metadata(0):
#> assays(1): raw
#> rownames: NULL
#> rowData names(14): in_ref gene_id_start ... strand_junction type
#> colnames(5): count_1 count_2 gtex_54020 gtex_51495 gtex_56930
#> colData names(2): samp_id case_control