For each junction, obtains the regions to obtain coverage

junction_cov_norm obtains regions of interest for each junction where coverage disruptions would be expected. If ends of junctions are annotated, it will use the overlapping exon definitions, picking the shortest exon when multiple overlap one end. If unannotated, will use a user-defined width (i.e. 20bp). To compare between samples (case/controls) coverage is normalised to a set region. By default, the boundaries of each gene associated to a junction are used as the region to normalise to.

junction_cov_norm(
  junctions,
  ref,
  unannot_width = 20,
  cov_paths_case,
  cov_paths_control,
  cov_chr_control = NULL
)

Arguments

junctions
junction data as a RangedSummarizedExperiment-class object.

ref
either path to gtf/gff3 or object of class TxDb imported using makeTxDbFromGFF.

unannot_width
integer scalar determining the width of the region to obtain coverage from when the end of of a junction does not overlap an existing exon.

cov_paths_case
paths to the BigWig/BAM files containing the coverage of your case samples. Must be the same length and order to the rows of the assays in junctions.

cov_paths_control
paths to the BigWig/BAM files

cov_chr_control
either "chr" or "no_chr", indicating the chromosome format of control coverage data. To be used if you know the chromosome format of the control BigWig/BAM files is different to that of your junctions.

Value

list containing sublists, one for cases and the other controls. Each sublist contains 3 matrices, corresponding the coverage for each sample across the 2 exons and intron associated with every junction.

Examples

if (FALSE) {
# leave this as not run for now to save time for R CMD check
ref <- "ftp://ftp.ensembl.org/pub/release-100/gtf/homo_sapiens/Homo_sapiens.GRCh38.100.gtf.gz"
ref <- GenomicFeatures::makeTxDbFromGFF(ref)
junctions <- cov_norm(junctions_annot_example, ref)
junctions
}