Normalise junction counts

junction_norm normalises the junction counts for each sample. First, each junction is clustered by finding all other junctions that share with it an acceptor or donor position. Then, a proportion-spliced-in (PSI) is calculated for each junction by dividing the raw junction count by the total number of counts in it's associated cluster.

junction_norm(junctions)

Arguments

junctions
junction data as a RangedSummarizedExperiment-class object.

Value

junctions as a SummarizedExperiment object with an additional assay containing the normalised counts.

Examples

if (!exists("junctions_normed")) {
    junctions_normed <- junction_norm(junctions_example)
}

#> [1] "2020-08-06 21:07:56 - Clustering junctions..."
#> [1] "2020-08-06 21:07:57 - Normalising junction counts..."
#> [1] "2020-08-06 21:07:58 - done!"

junctions_normed

#> class: RangedSummarizedExperiment 
#> dim: 97671 5 
#> metadata(0):
#> assays(2): raw norm
#> rownames: NULL
#> rowData names(2): index clusters
#> colnames(5): count_1 count_2 gtex_54020 gtex_51495 gtex_56930
#> colData names(2): samp_id case_control