Score patient junctions by their abnormality

junction_score will use the counts contained within the "norm" assay to calculate a deviation of each patient junction from the expected distribution of control junction counts. The function used to calculate this abnormality score can be user-inputted or left as the default z-score. Junctions will also be labelled based on whether they are up-regulated (+1) or down-regulated (-1) with respect to controls junction and this information is stored in the assay "direction" for use in outlier_aggregate.

junction_score(junctions, score_func = .zscore, ...)

Arguments

junctions
junction data as a RangedSummarizedExperiment-class object.

score_func
function to score junctions by their abnormality. By default, will use a z-score but can be switched to a user-defined function. This function must take as input an x and y argument, containing case and control counts respectively. This must return a numeric vector equal to the length of x with elements corresponding to a abnormality of each junction.

...
additional arguments passed to score_func.

Value

junctions as a RangedSummarizedExperiment-class object filtered for only case samples with an additional assay containing junction abnormality scores.

See also

Examples

if (!exists("junctions_normed")) {
    junctions_normed <- junction_norm(junctions_example)
}

#> [1] "2020-10-03 14:10:23 - Clustering junctions..."
#> [1] "2020-10-03 14:10:25 - Normalising junction counts..."
#> [1] "2020-10-03 14:10:26 - done!"

if (!exists("junctions_scored")) {
    junctions_scored <- junction_score(junctions_normed)
}

#> [1] "2020-10-03 14:10:26 - Calculating the direction of change of junctions..."
#> [1] "2020-10-03 14:10:27 - Generating junction abnormality score..."
#> [1] "2020-10-03 14:10:30 - done!"

junctions_scored

#> class: RangedSummarizedExperiment 
#> dim: 97671 2 
#> metadata(0):
#> assays(4): raw norm direction score
#> rownames: NULL
#> rowData names(2): index clusters
#> colnames(2): count_1 count_2
#> colData names(2): samp_id case_control