Aggregate outlier scores from per junction to cluster-level

outlier_aggregate will aggregate the outlier scores into a cluster-level. It will then rank each cluster based on this aggregated score and annotate each cluster with it's associated gene and transcript.

outlier_aggregate(
  junctions,
  samp_id_col = "samp_id",
  bp_param = BiocParallel::SerialParam()
)

Arguments

junctions
junction data as a RangedSummarizedExperiment-class object.

samp_id_col
name of the column in the SummarizedExperiment that details the sample ids.

bp_param
a BiocParallelParam-class instance denoting whether to parallelise the calculating of outlier scores across samples.

Value

DataFrame with one row per cluster detailing each cluster's associated junctions, outlier scores, ranks and genes.

See also

Examples

if (.Platform$OS.type != "windows") {

    # tell reticulate to use the python3 install
    # if windows skip this step
    reticulate::use_python(Sys.which("python3"), required = TRUE)
}

# use Genomic state to load txdb (GENCODE v31)
ref <- GenomicState::GenomicStateHub(version = "31", genome = "hg38", filetype = "TxDb")[[1]]

#> snapshotDate(): 2020-10-14

#> loading from cache

junctions_processed <- junction_process(
    junctions_example,
    ref,
    count_thresh = c("raw" = 5),
    n_samp = c("raw" = 1),
    types = c("ambig_gene", "unannotated"),
)

#> [1] "# Filtering junctions -----------------------------------------------------"
#> [1] "2020-10-16 15:08:43 - Filtering junctions..."
#> [1] "2020-10-16 15:08:43 - by count..."
#> [1] "2020-10-16 15:08:43 - done!"
#> [1] "# Annotating junctions ----------------------------------------------------"
#> [1] "2020-10-16 15:08:43 - Obtaining co-ordinates of annotated exons and junctions from gtf/gff3..."
#> [1] "2020-10-16 15:08:56 - Getting junction annotation using overlapping exons..."
#> [1] "2020-10-16 15:08:56 - Tidying junction annotation..."
#> [1] "2020-10-16 15:08:56 - Deriving junction categories..."
#> [1] "2020-10-16 15:08:57 - done!"
#> [1] "# Filtering junctions -----------------------------------------------------"
#> [1] "2020-10-16 15:08:57 - Filtering junctions..."
#> [1] "2020-10-16 15:08:57 - by type..."
#> [1] "2020-10-16 15:08:57 - done!"
#> [1] "# Normalise junctions -----------------------------------------------------"
#> [1] "2020-10-16 15:08:57 - Clustering junctions..."
#> [1] "2020-10-16 15:08:57 - Normalising junction counts..."
#> [1] "2020-10-16 15:08:57 - done!"
#> [1] "# Score junctions ---------------------------------------------------------"
#> [1] "2020-10-16 15:08:57 - Calculating the direction of change of junctions..."
#> [1] "2020-10-16 15:08:57 - Generating junction abnormality score..."
#> [1] "2020-10-16 15:08:57 - done!"

# obtain path to example bw on recount2
url <- recount::download_study(
    project = "SRP012682",
    type = "samples",
    download = FALSE
)

bw_path <- dasper:::.file_cache(url[1])

junctions_w_coverage <- coverage_process(
    junctions_processed,
    ref,
    coverage_paths_case = rep(bw_path, 2),
    coverage_paths_control = rep(bw_path, 3)
)

#> [1] "# Loading and normalising coverage ---------------------------------------------"
#> [1] "2020-10-16 15:08:57 - Obtaining exonic and intronic regions to load coverage from..."
#> [1] "2020-10-16 15:08:57 - Obtaining regions to use to normalise coverage..."
#> [1] "2020-10-16 15:09:01 - Loading coverage..."
#> [1] "2020-10-16 15:09:07 - Normalising coverage..."
#> [1] "2020-10-16 15:09:07 - done!"
#> [1] "# Scoring coverage ---------------------------------------------"
#> [1] "2020-10-16 15:09:07 - Generating coverage abnormality score..."
#> [1] "2020-10-16 15:09:07 - Obtaining regions with greatest coverage dysruption..."
#> [1] "2020-10-16 15:09:07 - done!"

junctions_w_outliers <- outlier_detect(junctions_w_coverage)

#> [1] "2020-10-16 15:09:07 - generating outlier scores for sample 1/2"
#> [1] "2020-10-16 15:09:08 - fitting outlier detection model with parameters: behaviour=deprecated, bootstrap=FALSE, contamination=auto, max_features=1, max_samples=auto, n_estimators=100, n_jobs=NULL, random_state=NULL, verbose=0, warm_start=FALSE"
#> [1] "2020-10-16 15:09:08 - fitting outlier detection model with parameters: behaviour=deprecated, bootstrap=FALSE, contamination=auto, max_features=1, max_samples=auto, n_estimators=100, n_jobs=NULL, random_state=NULL, verbose=0, warm_start=FALSE"
#> [1] "2020-10-16 15:09:09 - generating outlier scores for sample 2/2"
#> [1] "2020-10-16 15:09:09 - fitting outlier detection model with parameters: behaviour=deprecated, bootstrap=FALSE, contamination=auto, max_features=1, max_samples=auto, n_estimators=100, n_jobs=NULL, random_state=NULL, verbose=0, warm_start=FALSE"
#> [1] "2020-10-16 15:09:09 - fitting outlier detection model with parameters: behaviour=deprecated, bootstrap=FALSE, contamination=auto, max_features=1, max_samples=auto, n_estimators=100, n_jobs=NULL, random_state=NULL, verbose=0, warm_start=FALSE"
#> [1] "2020-10-16 15:09:09 - done!"

outlier_scores <- outlier_aggregate(junctions_w_outliers)

#> [1] "2020-10-16 15:09:09 - Aggregating outlier scores to cluster level... "
#> [1] "2020-10-16 15:09:10 - Annotating clusters with gene details..."
#> [1] "2020-10-16 15:09:10 - done!"