Detecting outlier junctions

outlier_detect will use the features in assays named feature_names as input into an unsupervised outlier detection algorithm to score each junction based on how outlier-y it looks in relation to other junctions in the patient. The default expected score and coverage_score features can be calculated using the junction_process and coverage_process respectively.

outlier_detect(
  junctions,
  feature_names = c("score", "coverage_score"),
  bp_param = BiocParallel::SerialParam(),
  ...
)

Arguments

junctions
junction data as a RangedSummarizedExperiment-class object.

feature_names
names of assays in junctions that are to be used as input into the outlier detection model.

bp_param
a BiocParallelParam-class instance denoting whether to parallelise the calculating of outlier scores across samples.

...
additional arguments passed to the outlier detection model (isolation forest) for setting parameters.

Value

junctions as a RangedSummarizedExperiment-class object with additional assays named outlier_score. The lower the outlier score, the more of an outlier a junction is observed to be.

See also