Wrapper for processing outliers

outlier_process wraps all "outlier_" prefixed functions in dasper. This is designed to simplify processing of the detecting outlier junctions for those familiar or uninterested with the intermediates.

outlier_process(
  junctions,
  feature_names = c("score", "coverage_score"),
  samp_id_col = "samp_id",
  bp_param = BiocParallel::SerialParam(),
  ...
)

Arguments

junctions
junction data as a RangedSummarizedExperiment-class object.

feature_names
names of assays in junctions that are to be used as input into the outlier detection model.

samp_id_col
name of the column in the SummarizedExperiment that details the sample ids.

bp_param
a BiocParallelParam-class instance denoting whether to parallelise the calculating of outlier scores across samples.

...
additional arguments passed to the outlier detection model (isolation forest) for setting parameters.

Value

DataFrame with one row per cluster detailing each cluster's associated junctions, outlier scores, ranks and genes.

See also

Examples

if (.Platform$OS.type != "windows") {
    # tell reticulate to use the python3 install
    # if windows skip this step
    reticulate::use_python(Sys.which("python3"), required = TRUE)
}

if (!exists("ref")) {
    # use Genomic state to load txdb (GENCODE v31)
    ref <- GenomicState::GenomicStateHub(version = "31", genome = "hg38", filetype = "TxDb")[[1]]
    # convert seqlevels to match junctions
    seqlevels(ref) <- stringr::str_replace(seqlevels(ref), "chr", "")
}

#> snapshotDate(): 2020-09-03

#> loading from cache

if (!exists("junctions_processed")) {
    junctions_processed <-
        junction_process(
            junctions_example,
            ref,
            count_thresh = c("raw" = 5),
            n_samp = c("raw" = 1),
            width_range = c(25, 1000000),
            types = c("ambig_gene", "unannotated"),
        )
    GenomeInfoDb::seqlevels(junctions_processed) <-
        paste0("chr", GenomeInfoDb::seqlevels(junctions_processed))
}

#> [1] "# Filtering junctions -----------------------------------------------------"
#> [1] "2020-10-14 17:57:58 - Filtering junctions..."
#> [1] "2020-10-14 17:57:58 - by count..."
#> [1] "2020-10-14 17:57:58 - by width..."
#> [1] "2020-10-14 17:57:58 - done!"
#> [1] "# Annotating junctions ----------------------------------------------------"
#> [1] "2020-10-14 17:57:58 - Obtaining co-ordinates of annotated exons and junctions from gtf/gff3..."
#> [1] "2020-10-14 17:58:15 - Getting junction annotation using overlapping exons..."
#> [1] "2020-10-14 17:58:15 - Tidying junction annotation..."
#> [1] "2020-10-14 17:58:16 - Deriving junction categories..."
#> [1] "2020-10-14 17:58:16 - done!"
#> [1] "# Filtering junctions -----------------------------------------------------"
#> [1] "2020-10-14 17:58:16 - Filtering junctions..."
#> [1] "2020-10-14 17:58:16 - by type..."
#> [1] "2020-10-14 17:58:16 - done!"
#> [1] "# Normalise junctions -----------------------------------------------------"
#> [1] "2020-10-14 17:58:16 - Clustering junctions..."
#> [1] "2020-10-14 17:58:16 - Normalising junction counts..."
#> [1] "2020-10-14 17:58:16 - done!"
#> [1] "# Score junctions ---------------------------------------------------------"
#> [1] "2020-10-14 17:58:16 - Calculating the direction of change of junctions..."
#> [1] "2020-10-14 17:58:16 - Generating junction abnormality score..."
#> [1] "2020-10-14 17:58:16 - done!"

# obtain path to example bw on recount2
url <- recount::download_study(
    project = "SRP012682",
    type = "samples",
    download = FALSE
)

bw_path <- dasper:::.file_cache(url[1])

if (!exists("junctions_w_coverage")) {
    junctions_w_coverage <-
        coverage_process(
            junctions_processed,
            ref,
            coverage_paths_case = rep(bw_path, 2),
            coverage_paths_control = rep(bw_path, 3)
        )
}

#> [1] "# Loading and normalising coverage ---------------------------------------------"
#> [1] "2020-10-14 17:58:17 - Obtaining exonic and intronic regions to load coverage from..."
#> [1] "2020-10-14 17:58:17 - Obtaining regions to use to normalise coverage..."

#> Warning: The 2 combined objects have no sequence levels in common. (Use
#>   suppressWarnings() to suppress this warning.)

#> [1] "2020-10-14 17:58:23 - Loading coverage..."

#> Warning: Total AUC across all regions was 0. Make sure chromsome format matches between input regions and bigWig/BAM file.

#> Warning: Total AUC across all regions was 0. Make sure chromsome format matches between input regions and bigWig/BAM file.

#> Warning: Total AUC across all regions was 0. Make sure chromsome format matches between input regions and bigWig/BAM file.

#> Warning: Total AUC across all regions was 0. Make sure chromsome format matches between input regions and bigWig/BAM file.

#> Warning: Total AUC across all regions was 0. Make sure chromsome format matches between input regions and bigWig/BAM file.

#> [1] "2020-10-14 17:58:26 - Normalising coverage..."
#> [1] "2020-10-14 17:58:26 - done!"
#> [1] "# Scoring coverage ---------------------------------------------"
#> [1] "2020-10-14 17:58:26 - Generating coverage abnormality score..."
#> [1] "2020-10-14 17:58:26 - Obtaining regions with greatest coverage dysruption..."
#> [1] "2020-10-14 17:58:27 - done!"

outlier_process(junctions_w_coverage)

#> [1] "# Detecting outliers using an isolation forest ------------------------------"
#> [1] "2020-10-14 17:58:27 - generating outlier scores for sample 1/2"
#> [1] "2020-10-14 17:58:27 - fitting outlier detection model with parameters: behaviour=deprecated, bootstrap=FALSE, contamination=auto, max_features=1, max_samples=auto, n_estimators=100, n_jobs=NULL, random_state=NULL, verbose=0, warm_start=FALSE"
#> [1] "2020-10-14 17:58:27 - fitting outlier detection model with parameters: behaviour=deprecated, bootstrap=FALSE, contamination=auto, max_features=1, max_samples=auto, n_estimators=100, n_jobs=NULL, random_state=NULL, verbose=0, warm_start=FALSE"
#> [1] "2020-10-14 17:58:27 - generating outlier scores for sample 2/2"
#> [1] "2020-10-14 17:58:27 - fitting outlier detection model with parameters: behaviour=deprecated, bootstrap=FALSE, contamination=auto, max_features=1, max_samples=auto, n_estimators=100, n_jobs=NULL, random_state=NULL, verbose=0, warm_start=FALSE"
#> [1] "2020-10-14 17:58:28 - fitting outlier detection model with parameters: behaviour=deprecated, bootstrap=FALSE, contamination=auto, max_features=1, max_samples=auto, n_estimators=100, n_jobs=NULL, random_state=NULL, verbose=0, warm_start=FALSE"
#> [1] "2020-10-14 17:58:28 - done!"
#> [1] "# Aggregating outlier scores to cluster-level -----------------------------"
#> [1] "2020-10-14 17:58:28 - Aggregating outlier scores to cluster level... "
#> [1] "2020-10-14 17:58:29 - Annotating clusters with gene details..."
#> [1] "2020-10-14 17:58:29 - done!"

#> DataFrame with 171 rows and 6 columns
#>         samp_id cluster_index mean_outlier_score      rank    gene_id_cluster
#>     <character>   <character>          <numeric> <numeric>    <CharacterList>
#> 1        samp_1          3857          -0.181007         1  ENSG00000100418.8
#> 2        samp_1          3984          -0.173240         2 ENSG00000100243.21
#> 3        samp_1          1243          -0.171569         3 ENSG00000142156.14
#> 4        samp_1          3680          -0.145308         4 ENSG00000100372.15
#> 5        samp_1          3992          -0.143253         5 ENSG00000128274.17
#> ...         ...           ...                ...       ...                ...
#> 167      samp_2          3072          0.0694639        85 ENSG00000100320.23
#> 168      samp_2          3116          0.0694639        85 ENSG00000100345.21
#> 169      samp_2          3119          0.0694639        85 ENSG00000100345.21
#> 170      samp_2           392          0.0694639        85 ENSG00000159128.14
#> 171      samp_2          4137          0.0694639        85 ENSG00000100376.12
#>                                             junctions
#>                                                <list>
#> 1     3857: 1:-0.120802:...,3858:-1:-0.241211:...,...
#> 2     3984: 1:-0.177244:...,3985:-1:-0.169236:...,...
#> 3     1244:-1:-0.208329:...,1245: 1:-0.134809:...,...
#> 4   3680:-1:-0.2302716:...,3681: 1:-0.0603451:...,...
#> 5   3992: 1:-0.0626296:...,3993:-1:-0.2238766:...,...
#> ...                                               ...
#> 167   3072:-1:0.0645729:...,3073: 1:0.0743549:...,...
#> 168   3116: 1:0.0743549:...,3117:-1:0.0645729:...,...
#> 169   3119: 1:0.0743549:...,3120:-1:0.0645729:...,...
#> 170     392:-1:0.0645729:...,393: 1:0.0743549:...,...
#> 171   4137:-1:0.0645729:...,4138: 1:0.0743549:...,...