Wrapper for processing coverage

coverage_process wraps all "coverage_" prefixed functions in dasper. This is designed to simplify processing of the coverage data for those familiar or uninterested with the intermediates.

coverage_process(
  junctions,
  ref,
  unannot_width = 20,
  coverage_paths_case,
  coverage_paths_control,
  coverage_chr_control = NULL,
  load_func = .coverage_load,
  bp_param = BiocParallel::SerialParam(),
  norm_const = 1,
  score_func = .zscore,
  ...
)

Arguments

junctions
junction data as a RangedSummarizedExperiment-class object.

ref
either path to gtf/gff3 or object of class TxDb-class.

unannot_width
integer scalar determining the width of the region to obtain coverage from when the end of of a junction does not overlap an existing exon.

coverage_paths_case
paths to the BigWig files containing the coverage of your case samples. Must be the same length and order to the samples in junctions.

coverage_paths_control
paths to the BigWig files

coverage_chr_control
either "chr" or "no_chr", indicating the chromosome format of control coverage data. Only used if you know the chromosome format of the control BigWig files is different to that of your junctions.

load_func
a function to use to load coverage. Currently only for internal use to increase testing speed.

bp_param
a BiocParallelParam-class instance denoting whether to parallelise the loading of coverage across BigWig files.

norm_const
numeric scaler to add to the normalisation coverage to avoid dividing by 0s and resulting NaN or Inf values.

score_func
function to score junctions by their abnormality. By default, will use a z-score but can be switched to a user-defined function. This function must take as input an x and y argument, containing case and control counts respectively. This must return a numeric vector equal to the length of x with elements corresponding to a abnormality of each junction.

...
additional arguments passed to score_func.

Value

RangedSummarizedExperiment-class object containing junction data with coverage scores.

See also

Examples

if (FALSE) {
ref <-
    "ftp://ftp.ensembl.org/pub/release-100/gtf/homo_sapiens/Homo_sapiens.GRCh38.100.gtf.gz"
junctions_w_coverage_2 <-
    coverage_process(
        junctions_example[, 1:2],
        ref,
        unannot_width = 20,
        coverage_paths_case,
        coverage_paths_control,
        coverage_chr_control = NULL,
        load_func = load_rand,
        norm_const = 2,
        score_func = .zscore,
        sd_const = 0.02
    )
junctions_w_coverage_2
}