Wrapper for processing coverage

coverage_process wraps all "coverage_" prefixed functions in dasper. This is designed to simplify processing of the coverage data for those familiar or uninterested with the intermediates.

coverage_process(
  junctions,
  ref,
  unannot_width = 20,
  coverage_paths_case,
  coverage_paths_control,
  coverage_chr_control = NULL,
  load_func = .coverage_load,
  bp_param = BiocParallel::SerialParam(),
  norm_const = 1,
  score_func = .zscore,
  ...
)

Arguments

junctions
junction data as a RangedSummarizedExperiment-class object.

ref
either path to gtf/gff3 or object of class TxDb-class.

unannot_width
integer scalar determining the width of the region to obtain coverage from when the end of of a junction does not overlap an existing exon.

coverage_paths_case
paths to the BigWig files containing the coverage of your case samples. Must be the same length and order to the samples in junctions.

coverage_paths_control
paths to the BigWig files

coverage_chr_control
either "chr" or "no_chr", indicating the chromosome format of control coverage data. Only used if you know the chromosome format of the control BigWig files is different to that of your junctions.

load_func
a function to use to load coverage. Currently only for internal use to increase testing speed.

bp_param
a BiocParallelParam-class instance denoting whether to parallelise the loading of coverage across BigWig files.

norm_const
numeric scaler to add to the normalisation coverage to avoid dividing by 0s and resulting NaN or Inf values.

score_func
function to score junctions by their abnormality. By default, will use a z-score but can be switched to a user-defined function. This function must take as input an x and y argument, containing case and control counts respectively. This must return a numeric vector equal to the length of x with elements corresponding to a abnormality of each junction.

...
additional arguments passed to score_func.

Value

RangedSummarizedExperiment-class object containing junction data with coverage scores.

See also

Examples

if (!exists("ref")) {
    # use Genomic state to load txdb (GENCODE v31)
    ref <- GenomicState::GenomicStateHub(version = "31", genome = "hg38", filetype = "TxDb")[[1]]
    # convert seqlevels to match junctions
    seqlevels(ref) <- stringr::str_replace(seqlevels(ref), "chr", "")
}

#> snapshotDate(): 2020-09-03

#> loading from cache

if (!exists("junctions_processed")) {
    junctions_processed <-
        junction_process(
            junctions_example,
            ref,
            count_thresh = c("raw" = 5),
            n_samp = c("raw" = 1),
            width_range = c(25, 1000000),
            types = c("ambig_gene", "unannotated"),
        )
    GenomeInfoDb::seqlevels(junctions_processed) <-
        paste0("chr", GenomeInfoDb::seqlevels(junctions_processed))
}

#> [1] "# Filtering junctions -----------------------------------------------------"
#> [1] "2020-10-14 17:54:22 - Filtering junctions..."
#> [1] "2020-10-14 17:54:22 - by count..."
#> [1] "2020-10-14 17:54:23 - by width..."
#> [1] "2020-10-14 17:54:23 - done!"
#> [1] "# Annotating junctions ----------------------------------------------------"
#> [1] "2020-10-14 17:54:23 - Obtaining co-ordinates of annotated exons and junctions from gtf/gff3..."
#> [1] "2020-10-14 17:54:41 - Getting junction annotation using overlapping exons..."
#> [1] "2020-10-14 17:54:43 - Tidying junction annotation..."
#> [1] "2020-10-14 17:54:43 - Deriving junction categories..."
#> [1] "2020-10-14 17:54:44 - done!"
#> [1] "# Filtering junctions -----------------------------------------------------"
#> [1] "2020-10-14 17:54:44 - Filtering junctions..."
#> [1] "2020-10-14 17:54:44 - by type..."
#> [1] "2020-10-14 17:54:44 - done!"
#> [1] "# Normalise junctions -----------------------------------------------------"
#> [1] "2020-10-14 17:54:44 - Clustering junctions..."
#> [1] "2020-10-14 17:54:44 - Normalising junction counts..."
#> [1] "2020-10-14 17:54:44 - done!"
#> [1] "# Score junctions ---------------------------------------------------------"
#> [1] "2020-10-14 17:54:44 - Calculating the direction of change of junctions..."
#> [1] "2020-10-14 17:54:44 - Generating junction abnormality score..."
#> [1] "2020-10-14 17:54:44 - done!"

# obtain path to example bw on recount2
url <- recount::download_study(
    project = "SRP012682",
    type = "samples",
    download = FALSE
)

bw_path <- dasper:::.file_cache(url[1])

if (!exists("junctions_w_coverage")) {
    junctions_w_coverage <-
        coverage_process(
            junctions_processed,
            ref,
            coverage_paths_case = rep(bw_path, 2),
            coverage_paths_control = rep(bw_path, 3)
        )
}

#> [1] "# Loading and normalising coverage ---------------------------------------------"
#> [1] "2020-10-14 17:54:45 - Obtaining exonic and intronic regions to load coverage from..."
#> [1] "2020-10-14 17:54:45 - Obtaining regions to use to normalise coverage..."

#> Warning: The 2 combined objects have no sequence levels in common. (Use
#>   suppressWarnings() to suppress this warning.)

#> [1] "2020-10-14 17:54:50 - Loading coverage..."

#> Warning: Total AUC across all regions was 0. Make sure chromsome format matches between input regions and bigWig/BAM file.

#> Warning: Total AUC across all regions was 0. Make sure chromsome format matches between input regions and bigWig/BAM file.

#> Warning: Total AUC across all regions was 0. Make sure chromsome format matches between input regions and bigWig/BAM file.

#> Warning: Total AUC across all regions was 0. Make sure chromsome format matches between input regions and bigWig/BAM file.

#> Warning: Total AUC across all regions was 0. Make sure chromsome format matches between input regions and bigWig/BAM file.

#> [1] "2020-10-14 17:54:53 - Normalising coverage..."
#> [1] "2020-10-14 17:54:53 - done!"
#> [1] "# Scoring coverage ---------------------------------------------"
#> [1] "2020-10-14 17:54:53 - Generating coverage abnormality score..."
#> [1] "2020-10-14 17:54:53 - Obtaining regions with greatest coverage dysruption..."
#> [1] "2020-10-14 17:54:54 - done!"

junctions_w_coverage

#> class: RangedSummarizedExperiment 
#> dim: 4705 2 
#> metadata(0):
#> assays(6): raw norm ... coverage_region coverage_score
#> rownames: NULL
#> rowData names(16): in_ref gene_id_start ... index clusters
#> colnames(2): count_1 count_2
#> colData names(2): samp_id case_control