Score coverage by their abnormality

coverage_score will score disruptions in the coverage across the intronic/exonic regions associated with each junction. This abnormality score generated by score_func operates by calculating the deviation of the coverage in patients to a coverage across the same regions in controls. Then, for each junction it obtains the score of the region with the greatest disruption.

coverage_score(junctions, coverage, score_func = .zscore, ...)

Arguments

junctions
junction data as a RangedSummarizedExperiment-class object.

coverage
list containing normalised coverage data that is outputted from coverage_norm.

score_func
function to score junctions by their abnormality. By default, will use a z-score but can be switched to a user-defined function. This function must take as input an x and y argument, containing case and control counts respectively. This must return a numeric vector equal to the length of x with elements corresponding to a abnormality of each junction.

...
additional arguments passed to score_func.

Value

junctions as SummarizedExperiment object with additional assays named "coverage_region" and "coverage_score". "coverage_region" labels the region of greatest disruption (1 = exon_start, 2 = exon_end, 3 = intron) and "coverage_score" contains the abnormality scores of the region with the greatest disruption.

See also

Examples

if (!exists("ref")) {
    ref <- "ftp://ftp.ensembl.org/pub/release-100/gtf/homo_sapiens/Homo_sapiens.GRCh38.100.gtf.gz"
    ref <- GenomicFeatures::makeTxDbFromGFF(ref)
}

#> Import genomic features from the file as a GRanges object ... 

#> OK

#> Prepare the 'metadata' data frame ... 

#> OK

#> Make the TxDb object ... 

#> Warning: The "phase" metadata column contains non-NA values for features of type
#>   stop_codon. This information was ignored.

#> OK

if (!exists("junctions_processed")) {
    junctions_processed <-
        junction_process(
            junctions_example,
            ref,
            count_thresh = c("raw" = 5),
            n_samp = c("raw" = 1),
            width_range = c(25, 1000000),
            types = c("ambig_gene", "unannotated"),
        )
    GenomeInfoDb::seqlevels(junctions_processed) <-
        paste0("chr", GenomeInfoDb::seqlevels(junctions_processed))
}

#> [1] "# Filtering junctions -----------------------------------------------------"
#> [1] "2020-10-14 14:36:23 - Filtering junctions..."
#> [1] "2020-10-14 14:36:23 - by count..."
#> [1] "2020-10-14 14:36:23 - by width..."
#> [1] "2020-10-14 14:36:23 - done!"
#> [1] "# Annotating junctions ----------------------------------------------------"
#> [1] "2020-10-14 14:36:23 - Obtaining co-ordinates of annotated exons and junctions from gtf/gff3..."
#> [1] "2020-10-14 14:36:39 - Getting junction annotation using overlapping exons..."
#> [1] "2020-10-14 14:36:39 - Tidying junction annotation..."
#> [1] "2020-10-14 14:36:39 - Deriving junction categories..."
#> [1] "2020-10-14 14:36:39 - done!"
#> [1] "# Filtering junctions -----------------------------------------------------"
#> [1] "2020-10-14 14:36:39 - Filtering junctions..."
#> [1] "2020-10-14 14:36:39 - by type..."
#> [1] "2020-10-14 14:36:39 - done!"
#> [1] "# Normalise junctions -----------------------------------------------------"
#> [1] "2020-10-14 14:36:39 - Clustering junctions..."
#> [1] "2020-10-14 14:36:40 - Normalising junction counts..."
#> [1] "2020-10-14 14:36:40 - done!"
#> [1] "# Score junctions ---------------------------------------------------------"
#> [1] "2020-10-14 14:36:40 - Calculating the direction of change of junctions..."
#> [1] "2020-10-14 14:36:40 - Generating junction abnormality score..."
#> [1] "2020-10-14 14:36:40 - done!"

# obtain path to example bw on recount2
url <- recount::download_study(
    project = "SRP012682",
    type = "samples",
    download = FALSE
)

bw_path <- file.path(tempdir(), basename(url[1]))

if (!file.exists(bw_path)) {
    download.file(url[1], bw_path)
}

if (!exists("coverage")) {
    coverage <- coverage_norm(
        junctions_processed,
        ref,
        coverage_paths_case = rep(bw_path, 2),
        coverage_paths_control = rep(bw_path, 3),
        bp_param = BiocParallel::MulticoreParam(5)
    )
}

#> [1] "2020-10-14 14:36:40 - Obtaining exonic and intronic regions to load coverage from..."
#> [1] "2020-10-14 14:36:40 - Obtaining regions to use to normalise coverage..."

#> Warning: The 2 combined objects have no sequence levels in common. (Use
#>   suppressWarnings() to suppress this warning.)

#> [1] "2020-10-14 14:36:45 - Loading coverage..."
#> [1] "2020-10-14 14:36:50 - Normalising coverage..."
#> [1] "2020-10-14 14:36:50 - done!"

if (!exists("junctions_cov_scored")) {
    junctions_cov_scored <- coverage_score(junctions_processed, coverage)
}

#> [1] "2020-10-14 14:36:50 - Generating coverage abnormality score..."
#> [1] "2020-10-14 14:36:50 - Obtaining regions with greatest coverage dysruption..."
#> [1] "2020-10-14 14:36:50 - done!"