Wrapper function for processing junctions

junction_process wraps all "junction_" prefixed functions in dasper except junction_load. This is designed to simplify processing of the junction data for those familiar or uninterested with the intermediates.

junction_process(
  junctions,
  ref,
  count_thresh = c(raw = 5),
  n_samp = c(raw = 1),
  width_range = NULL,
  types = NULL,
  regions = NULL,
  score_func = .zscore,
  ...
)

Arguments

junctions
junction data as a RangedSummarizedExperiment-class object.

ref
either path to gtf/gff3 or object of class TxDb imported using makeTxDbFromGFF.

count_thresh
named vector with names matching the names of the assays in junctions. Values denote the number of counts below which a junction will be filtered out.

n_samp
named vector with names matching the names of the assays in junctions. Values denotes number of samples that have to express the junction above the count_thresh in order for that junction to not be filtered.

width_range
numeric vector of length 2. The first element denoting the lower limit of junction width and the second the upper limit. Junctions with widths outside this range will be filtered out.

types
any junctions matching these types, derived form junction_annot e.g. "unannotated" will be filtered out.

regions
any junctions overlapping this set of regions (in a GRanges format) will be filtered out.

score_func
function to score junctions by abnromality. By default, will use a z-score. But this can be switched dependent on the format of the user's junction data. Function must take as input an x and y argument. x being the case junctions and y being the controls. This must return a numeric vector equal to the length of x with each element corresponding to a abnormality of a junction in each case sample.

...
additional arguments passed to score_func.

Value

RangedSummarizedExperiment-class object containing filtered, annotated, normalised junction data with scores.

Examples

ref <-
    "ftp://ftp.ensembl.org/pub/release-100/gtf/homo_sapiens/Homo_sapiens.GRCh38.100.gtf.gz"
junctions_processed <-
    junction_process(
        junctions_example,
        ref,
        count_thresh = c("raw" = 5),
        n_samp = c("raw" = 1),
        width_range = c(25, 1000000),
        types = c("ambig_gene", "unannotated"),
    )

#> [1] "# Filtering junctions -----------------------------------------------------"
#> [1] "2020-08-06 10:24:43 - Filtering junctions..."
#> [1] "2020-08-06 10:24:43 - by count..."
#> [1] "2020-08-06 10:24:43 - by width..."
#> [1] "2020-08-06 10:24:43 - done!"
#> [1] "# Annotating junctions ----------------------------------------------------"
#> [1] "2020-08-06 10:24:43 - Obtaining co-ordinates of annotated exons and junctions from gtf/gff3..."
#> [1] "2020-08-06 10:24:43 - Importing gtf/gff3 as a TxDb..."

#> Import genomic features from the file as a GRanges object ... 

#> OK

#> Prepare the 'metadata' data frame ... 

#> OK

#> Make the TxDb object ... 

#> Warning: The "phase" metadata column contains non-NA values for features of type
#>   stop_codon. This information was ignored.

#> OK

#> [1] "2020-08-06 10:26:00 - Getting junction annotation using overlapping exons..."
#> [1] "2020-08-06 10:26:00 - Tidying junction annotation..."
#> [1] "2020-08-06 10:26:01 - Deriving junction categories..."
#> [1] "2020-08-06 10:26:01 - done!"
#> [1] "# Filtering junctions -----------------------------------------------------"
#> [1] "2020-08-06 10:26:01 - Filtering junctions..."
#> [1] "2020-08-06 10:26:01 - by type..."
#> [1] "2020-08-06 10:26:01 - done!"
#> [1] "# Normalise junctions -----------------------------------------------------"
#> [1] "2020-08-06 10:26:01 - Clustering junctions..."
#> [1] "2020-08-06 10:26:01 - Normalising junction counts..."
#> [1] "2020-08-06 10:26:01 - done!"
#> [1] "# Score junctions ---------------------------------------------------------"
#> [1] "2020-08-06 10:26:01 - Generating junction abnormality score..."
#> [1] "2020-08-06 10:26:01 - done!"

junctions_processed

#> class: RangedSummarizedExperiment 
#> dim: 4706 2 
#> metadata(0):
#> assays(3): raw norm score
#> rownames: NULL
#> rowData names(16): in_ref gene_id_start ... index clusters
#> colnames(2): count_1 count_2
#> colData names(2): samp_id case_control