Aggregate outlier scores from per junction to cluster-level

outlier_aggregate will aggregate the outlier scores into a cluster-level. It will then rank each cluster based on this aggregated score and annotate each cluster with it's associated gene and transcript.

outlier_aggregate(
  junctions,
  samp_id_col = "samp_id",
  bp_param = BiocParallel::SerialParam()
)

Arguments

junctions
junction data as a RangedSummarizedExperiment-class object.

samp_id_col
name of the column in the SummarizedExperiment that details the sample ids.

bp_param
a BiocParallelParam-class instance denoting whether to parallelise the calculating of outlier scores across samples.

Value

DataFrame with one row per cluster detailing each cluster's associated junctions, outlier scores, ranks and genes.

See also

Examples

if (!exists("ref")) {
    # use Genomic state to load txdb (GENCODE v31)
    ref <- GenomicState::GenomicStateHub(version = "31", genome = "hg38", filetype = "TxDb")[[1]]
    # convert seqlevels to match junctions
    seqlevels(ref) <- stringr::str_replace(seqlevels(ref), "chr", "")
}

#> snapshotDate(): 2020-09-03

#> loading from cache

if (!exists("junctions_processed")) {
    junctions_processed <-
        junction_process(
            junctions_example,
            ref,
            count_thresh = c("raw" = 5),
            n_samp = c("raw" = 1),
            width_range = c(25, 1000000),
            types = c("ambig_gene", "unannotated"),
        )
    GenomeInfoDb::seqlevels(junctions_processed) <-
        paste0("chr", GenomeInfoDb::seqlevels(junctions_processed))
}

#> [1] "# Filtering junctions -----------------------------------------------------"
#> [1] "2020-10-14 17:56:44 - Filtering junctions..."
#> [1] "2020-10-14 17:56:44 - by count..."
#> [1] "2020-10-14 17:56:44 - by width..."
#> [1] "2020-10-14 17:56:44 - done!"
#> [1] "# Annotating junctions ----------------------------------------------------"
#> [1] "2020-10-14 17:56:44 - Obtaining co-ordinates of annotated exons and junctions from gtf/gff3..."
#> [1] "2020-10-14 17:57:00 - Getting junction annotation using overlapping exons..."
#> [1] "2020-10-14 17:57:01 - Tidying junction annotation..."
#> [1] "2020-10-14 17:57:01 - Deriving junction categories..."
#> [1] "2020-10-14 17:57:01 - done!"
#> [1] "# Filtering junctions -----------------------------------------------------"
#> [1] "2020-10-14 17:57:01 - Filtering junctions..."
#> [1] "2020-10-14 17:57:01 - by type..."
#> [1] "2020-10-14 17:57:01 - done!"
#> [1] "# Normalise junctions -----------------------------------------------------"
#> [1] "2020-10-14 17:57:02 - Clustering junctions..."
#> [1] "2020-10-14 17:57:02 - Normalising junction counts..."
#> [1] "2020-10-14 17:57:02 - done!"
#> [1] "# Score junctions ---------------------------------------------------------"
#> [1] "2020-10-14 17:57:02 - Calculating the direction of change of junctions..."
#> [1] "2020-10-14 17:57:02 - Generating junction abnormality score..."
#> [1] "2020-10-14 17:57:02 - done!"

# obtain path to example bw on recount2
url <- recount::download_study(
    project = "SRP012682",
    type = "samples",
    download = FALSE
)

bw_path <- dasper:::.file_cache(url[1])

if (!exists("junctions_w_coverage")) {
    junctions_w_coverage <-
        coverage_process(
            junctions_processed,
            ref,
            coverage_paths_case = rep(bw_path, 2),
            coverage_paths_control = rep(bw_path, 3)
        )
}

#> [1] "# Loading and normalising coverage ---------------------------------------------"
#> [1] "2020-10-14 17:57:02 - Obtaining exonic and intronic regions to load coverage from..."
#> [1] "2020-10-14 17:57:03 - Obtaining regions to use to normalise coverage..."

#> Warning: The 2 combined objects have no sequence levels in common. (Use
#>   suppressWarnings() to suppress this warning.)

#> [1] "2020-10-14 17:57:09 - Loading coverage..."

#> Warning: Total AUC across all regions was 0. Make sure chromsome format matches between input regions and bigWig/BAM file.

#> Warning: Total AUC across all regions was 0. Make sure chromsome format matches between input regions and bigWig/BAM file.

#> Warning: Total AUC across all regions was 0. Make sure chromsome format matches between input regions and bigWig/BAM file.

#> Warning: Total AUC across all regions was 0. Make sure chromsome format matches between input regions and bigWig/BAM file.

#> Warning: Total AUC across all regions was 0. Make sure chromsome format matches between input regions and bigWig/BAM file.

#> [1] "2020-10-14 17:57:12 - Normalising coverage..."
#> [1] "2020-10-14 17:57:12 - done!"
#> [1] "# Scoring coverage ---------------------------------------------"
#> [1] "2020-10-14 17:57:12 - Generating coverage abnormality score..."
#> [1] "2020-10-14 17:57:12 - Obtaining regions with greatest coverage dysruption..."
#> [1] "2020-10-14 17:57:12 - done!"

if (!exists("junctions_w_outliers")) {
    junctions_w_outliers <- outlier_detect(junctions_w_coverage)
}

#> [1] "2020-10-14 17:57:12 - generating outlier scores for sample 1/2"
#> [1] "2020-10-14 17:57:13 - fitting outlier detection model with parameters: behaviour=deprecated, bootstrap=FALSE, contamination=auto, max_features=1, max_samples=auto, n_estimators=100, n_jobs=NULL, random_state=NULL, verbose=0, warm_start=FALSE"
#> [1] "2020-10-14 17:57:14 - fitting outlier detection model with parameters: behaviour=deprecated, bootstrap=FALSE, contamination=auto, max_features=1, max_samples=auto, n_estimators=100, n_jobs=NULL, random_state=NULL, verbose=0, warm_start=FALSE"
#> [1] "2020-10-14 17:57:14 - generating outlier scores for sample 2/2"
#> [1] "2020-10-14 17:57:14 - fitting outlier detection model with parameters: behaviour=deprecated, bootstrap=FALSE, contamination=auto, max_features=1, max_samples=auto, n_estimators=100, n_jobs=NULL, random_state=NULL, verbose=0, warm_start=FALSE"
#> [1] "2020-10-14 17:57:15 - fitting outlier detection model with parameters: behaviour=deprecated, bootstrap=FALSE, contamination=auto, max_features=1, max_samples=auto, n_estimators=100, n_jobs=NULL, random_state=NULL, verbose=0, warm_start=FALSE"
#> [1] "2020-10-14 17:57:15 - done!"

outlier_scores <- outlier_aggregate(junctions_w_outliers)

#> [1] "2020-10-14 17:57:15 - Aggregating outlier scores to cluster level... "
#> [1] "2020-10-14 17:57:16 - Annotating clusters with gene details..."
#> [1] "2020-10-14 17:57:16 - done!"