Detecting outlier junctions

outlier_detect will use the features in assays named feature_names as input into an unsupervised outlier detection algorithm to score each junction based on how outlier-y it looks in relation to other junctions in the patient. The default expected score and coverage_score features can be calculated using the junction_process and coverage_process respectively.

outlier_detect(
  junctions,
  feature_names = c("score", "coverage_score"),
  bp_param = BiocParallel::SerialParam(),
  ...
)

Arguments

junctions
junction data as a RangedSummarizedExperiment-class object.

feature_names
names of assays in junctions that are to be used as input into the outlier detection model.

bp_param
a BiocParallelParam-class instance denoting whether to parallelise the calculating of outlier scores across samples.

...
additional arguments passed to the outlier detection model (isolation forest) for setting parameters.

Value

junctions as a RangedSummarizedExperiment-class object with additional assays named outlier_score. The lower the outlier score, the more of an outlier a junction is observed to be.

See also

Examples

if (.Platform$OS.type != "windows") {

    # tell reticulate to use the python3 install
    # if windows skip this step
    reticulate::use_python(Sys.which("python3"), required = TRUE)
}

if (!exists("ref")) {
    ref <- "ftp://ftp.ensembl.org/pub/release-100/gtf/homo_sapiens/Homo_sapiens.GRCh38.100.gtf.gz"
    ref <- GenomicFeatures::makeTxDbFromGFF(ref)
}

#> Import genomic features from the file as a GRanges object ... 

#> OK

#> Prepare the 'metadata' data frame ... 

#> OK

#> Make the TxDb object ... 

#> Warning: The "phase" metadata column contains non-NA values for features of type
#>   stop_codon. This information was ignored.

#> OK

if (!exists("junctions_processed")) {
    junctions_processed <-
        junction_process(
            junctions_example,
            ref,
            count_thresh = c("raw" = 5),
            n_samp = c("raw" = 1),
            width_range = c(25, 1000000),
            types = c("ambig_gene", "unannotated"),
        )
    GenomeInfoDb::seqlevels(junctions_processed) <-
        paste0("chr", GenomeInfoDb::seqlevels(junctions_processed))
}

#> [1] "# Filtering junctions -----------------------------------------------------"
#> [1] "2020-10-02 23:00:11 - Filtering junctions..."
#> [1] "2020-10-02 23:00:11 - by count..."
#> [1] "2020-10-02 23:00:11 - by width..."
#> [1] "2020-10-02 23:00:11 - done!"
#> [1] "# Annotating junctions ----------------------------------------------------"
#> [1] "2020-10-02 23:00:11 - Obtaining co-ordinates of annotated exons and junctions from gtf/gff3..."
#> [1] "2020-10-02 23:00:26 - Getting junction annotation using overlapping exons..."
#> [1] "2020-10-02 23:00:26 - Tidying junction annotation..."
#> [1] "2020-10-02 23:00:26 - Deriving junction categories..."
#> [1] "2020-10-02 23:00:26 - done!"
#> [1] "# Filtering junctions -----------------------------------------------------"
#> [1] "2020-10-02 23:00:26 - Filtering junctions..."
#> [1] "2020-10-02 23:00:26 - by type..."
#> [1] "2020-10-02 23:00:26 - done!"
#> [1] "# Normalise junctions -----------------------------------------------------"
#> [1] "2020-10-02 23:00:26 - Clustering junctions..."
#> [1] "2020-10-02 23:00:27 - Normalising junction counts..."
#> [1] "2020-10-02 23:00:27 - done!"
#> [1] "# Score junctions ---------------------------------------------------------"
#> [1] "2020-10-02 23:00:27 - Calculating the direction of change of junctions..."
#> [1] "2020-10-02 23:00:27 - Generating junction abnormality score..."
#> [1] "2020-10-02 23:00:27 - done!"

url <- recount::download_study(
    project = "SRP012682",
    type = "samples",
    download = FALSE
)
bw_path <- file.path(tempdir(), basename(url[1]))

if (!file.exists(bw_path)) {
    download.file(url[1], bw_path)
}

if (!exists("junctions_w_coverage")) {
    junctions_w_coverage <-
        coverage_process(
            junctions_processed,
            ref,
            coverage_paths_case = rep(bw_path, 2),
            coverage_paths_control = rep(bw_path, 3)
        )
}

#> [1] "# Loading and normalising coverage ---------------------------------------------"
#> [1] "2020-10-02 23:00:27 - Obtaining exonic and intronic regions to load coverage from..."
#> [1] "2020-10-02 23:00:27 - Obtaining regions to use to normalise coverage..."

#> Warning: The 2 combined objects have no sequence levels in common. (Use
#>   suppressWarnings() to suppress this warning.)

#> [1] "2020-10-02 23:00:32 - Loading coverage..."

#> Warning: Total AUC across all regions was 0. Make sure chromsome format matches between input regions and bigWig/BAM file.

#> Warning: Total AUC across all regions was 0. Make sure chromsome format matches between input regions and bigWig/BAM file.

#> Warning: Total AUC across all regions was 0. Make sure chromsome format matches between input regions and bigWig/BAM file.

#> Warning: Total AUC across all regions was 0. Make sure chromsome format matches between input regions and bigWig/BAM file.

#> Warning: Total AUC across all regions was 0. Make sure chromsome format matches between input regions and bigWig/BAM file.

#> [1] "2020-10-02 23:00:34 - Normalising coverage..."
#> [1] "2020-10-02 23:00:34 - done!"
#> [1] "# Scoring coverage ---------------------------------------------"
#> [1] "2020-10-02 23:00:34 - Generating coverage abnormality score..."
#> [1] "2020-10-02 23:00:35 - Obtaining regions with greatest coverage dysruption..."
#> [1] "2020-10-02 23:00:35 - done!"

if (!exists("junctions_w_outliers")) {
    junctions_w_outliers <- outlier_detect(junctions_w_coverage)
}

#> [1] "2020-10-02 23:00:35 - generating outlier scores for sample 1/2"
#> [1] "2020-10-02 23:00:35 - fitting outlier detection model with parameters: behaviour=deprecated, bootstrap=FALSE, contamination=auto, max_features=1, max_samples=auto, n_estimators=100, n_jobs=NULL, random_state=NULL, verbose=0, warm_start=FALSE"
#> [1] "2020-10-02 23:00:35 - fitting outlier detection model with parameters: behaviour=deprecated, bootstrap=FALSE, contamination=auto, max_features=1, max_samples=auto, n_estimators=100, n_jobs=NULL, random_state=NULL, verbose=0, warm_start=FALSE"
#> [1] "2020-10-02 23:00:35 - generating outlier scores for sample 2/2"
#> [1] "2020-10-02 23:00:35 - fitting outlier detection model with parameters: behaviour=deprecated, bootstrap=FALSE, contamination=auto, max_features=1, max_samples=auto, n_estimators=100, n_jobs=NULL, random_state=NULL, verbose=0, warm_start=FALSE"
#> [1] "2020-10-02 23:00:36 - fitting outlier detection model with parameters: behaviour=deprecated, bootstrap=FALSE, contamination=auto, max_features=1, max_samples=auto, n_estimators=100, n_jobs=NULL, random_state=NULL, verbose=0, warm_start=FALSE"
#> [1] "2020-10-02 23:00:36 - done!"

junctions_w_outliers

#> class: RangedSummarizedExperiment 
#> dim: 4706 2 
#> metadata(0):
#> assays(7): raw norm ... coverage_score outlier_score
#> rownames: NULL
#> rowData names(16): in_ref gene_id_start ... index clusters
#> colnames(2): count_1 count_2
#> colData names(2): samp_id case_control