Wrapper for processing outliers

outlier_process wraps all "outlier_" prefixed functions in dasper. This is designed to simplify processing of the detecting outlier junctions for those familiar or uninterested with the intermediates.

outlier_process(
  junctions,
  feature_names = c("score", "coverage_score"),
  samp_id_col = "samp_id",
  bp_param = BiocParallel::SerialParam(),
  ...
)

Arguments

junctions
junction data as a RangedSummarizedExperiment-class object.

feature_names
names of assays in junctions that are to be used as input into the outlier detection model.

samp_id_col
name of the column in the SummarizedExperiment that details the sample ids.

bp_param
a BiocParallelParam-class instance denoting whether to parallelise the calculating of outlier scores across samples.

...
additional arguments passed to the outlier detection model (isolation forest) for setting parameters.

Value

DataFrame with one row per cluster detailing each cluster's associated junctions, outlier scores, ranks and genes.

See also

Examples

if (.Platform$OS.type != "windows") {
    # tell reticulate to use the python3 install
    # if windows skip this step
    reticulate::use_python(Sys.which("python3"), required = TRUE)
}

# use Genomic state to load txdb (GENCODE v31)
ref <- GenomicState::GenomicStateHub(version = "31", genome = "hg38", filetype = "TxDb")[[1]]

#> snapshotDate(): 2020-10-14

#> loading from cache

junctions_processed <- junction_process(
    junctions_example,
    ref,
    count_thresh = c("raw" = 5),
    n_samp = c("raw" = 1),
    types = c("ambig_gene", "unannotated"),
)

#> [1] "# Filtering junctions -----------------------------------------------------"
#> [1] "2020-10-16 14:59:23 - Filtering junctions..."
#> [1] "2020-10-16 14:59:23 - by count..."
#> [1] "2020-10-16 14:59:23 - done!"
#> [1] "# Annotating junctions ----------------------------------------------------"
#> [1] "2020-10-16 14:59:23 - Obtaining co-ordinates of annotated exons and junctions from gtf/gff3..."
#> [1] "2020-10-16 14:59:38 - Getting junction annotation using overlapping exons..."
#> [1] "2020-10-16 14:59:39 - Tidying junction annotation..."
#> [1] "2020-10-16 14:59:39 - Deriving junction categories..."
#> [1] "2020-10-16 14:59:39 - done!"
#> [1] "# Filtering junctions -----------------------------------------------------"
#> [1] "2020-10-16 14:59:39 - Filtering junctions..."
#> [1] "2020-10-16 14:59:39 - by type..."
#> [1] "2020-10-16 14:59:39 - done!"
#> [1] "# Normalise junctions -----------------------------------------------------"
#> [1] "2020-10-16 14:59:39 - Clustering junctions..."
#> [1] "2020-10-16 14:59:39 - Normalising junction counts..."
#> [1] "2020-10-16 14:59:39 - done!"
#> [1] "# Score junctions ---------------------------------------------------------"
#> [1] "2020-10-16 14:59:39 - Calculating the direction of change of junctions..."
#> [1] "2020-10-16 14:59:39 - Generating junction abnormality score..."
#> [1] "2020-10-16 14:59:39 - done!"

# obtain path to example bw on recount2
url <- recount::download_study(
    project = "SRP012682",
    type = "samples",
    download = FALSE
)

bw_path <- dasper:::.file_cache(url[1])

junctions_w_coverage <- coverage_process(
    junctions_processed,
    ref,
    coverage_paths_case = rep(bw_path, 2),
    coverage_paths_control = rep(bw_path, 3)
)

#> [1] "# Loading and normalising coverage ---------------------------------------------"
#> [1] "2020-10-16 14:59:40 - Obtaining exonic and intronic regions to load coverage from..."
#> [1] "2020-10-16 14:59:40 - Obtaining regions to use to normalise coverage..."
#> [1] "2020-10-16 14:59:44 - Loading coverage..."
#> [1] "2020-10-16 14:59:49 - Normalising coverage..."
#> [1] "2020-10-16 14:59:49 - done!"
#> [1] "# Scoring coverage ---------------------------------------------"
#> [1] "2020-10-16 14:59:49 - Generating coverage abnormality score..."
#> [1] "2020-10-16 14:59:50 - Obtaining regions with greatest coverage dysruption..."
#> [1] "2020-10-16 14:59:50 - done!"

outlier_process(junctions_w_coverage)

#> [1] "# Detecting outliers using an isolation forest ------------------------------"
#> [1] "2020-10-16 14:59:50 - generating outlier scores for sample 1/2"
#> [1] "2020-10-16 14:59:50 - fitting outlier detection model with parameters: behaviour=deprecated, bootstrap=FALSE, contamination=auto, max_features=1, max_samples=auto, n_estimators=100, n_jobs=NULL, random_state=NULL, verbose=0, warm_start=FALSE"
#> [1] "2020-10-16 14:59:50 - fitting outlier detection model with parameters: behaviour=deprecated, bootstrap=FALSE, contamination=auto, max_features=1, max_samples=auto, n_estimators=100, n_jobs=NULL, random_state=NULL, verbose=0, warm_start=FALSE"
#> [1] "2020-10-16 14:59:51 - generating outlier scores for sample 2/2"
#> [1] "2020-10-16 14:59:51 - fitting outlier detection model with parameters: behaviour=deprecated, bootstrap=FALSE, contamination=auto, max_features=1, max_samples=auto, n_estimators=100, n_jobs=NULL, random_state=NULL, verbose=0, warm_start=FALSE"
#> [1] "2020-10-16 14:59:51 - fitting outlier detection model with parameters: behaviour=deprecated, bootstrap=FALSE, contamination=auto, max_features=1, max_samples=auto, n_estimators=100, n_jobs=NULL, random_state=NULL, verbose=0, warm_start=FALSE"
#> [1] "2020-10-16 14:59:51 - done!"
#> [1] "# Aggregating outlier scores to cluster-level -----------------------------"
#> [1] "2020-10-16 14:59:51 - Aggregating outlier scores to cluster level... "
#> [1] "2020-10-16 14:59:52 - Annotating clusters with gene details..."
#> [1] "2020-10-16 14:59:52 - done!"

#> DataFrame with 170 rows and 6 columns
#>         samp_id cluster_index mean_outlier_score      rank    gene_id_cluster
#>     <character>   <character>          <numeric> <numeric>    <CharacterList>
#> 1        samp_1          3984          -0.190907         1 ENSG00000100243.21
#> 2        samp_1          1243          -0.184554         2 ENSG00000142156.14
#> 3        samp_1          3857          -0.178024         3  ENSG00000100418.8
#> 4        samp_1           968          -0.151837         4  ENSG00000160213.7
#> 5        samp_1          3680          -0.143322         5 ENSG00000100372.15
#> ...         ...           ...                ...       ...                ...
#> 166      samp_2          3072          0.0722241        85 ENSG00000100320.23
#> 167      samp_2          3116          0.0722241        85 ENSG00000100345.21
#> 168      samp_2          3119          0.0722241        85 ENSG00000100345.21
#> 169      samp_2           392          0.0722241        85 ENSG00000159128.14
#> 170      samp_2          4137          0.0722241        85 ENSG00000100376.12
#>                                           junctions
#>                                              <list>
#> 1   3984: 1:-0.176468:...,3985:-1:-0.205345:...,...
#> 2   1244:-1:-0.237611:...,1245: 1:-0.131496:...,...
#> 3   3857: 1:-0.116081:...,3858:-1:-0.239967:...,...
#> 4   968: 1:-0.0846473:...,969:-1:-0.2190258:...,...
#> 5   3680:-1:-0.232546:...,3681: 1:-0.054099:...,...
#> ...                                             ...
#> 166 3072:-1:0.0723692:...,3073: 1:0.0720790:...,...
#> 167 3116: 1:0.0720790:...,3117:-1:0.0723692:...,...
#> 168 3119: 1:0.0720790:...,3120:-1:0.0723692:...,...
#> 169   392:-1:0.0723692:...,393: 1:0.0720790:...,...
#> 170 4137:-1:0.0723692:...,4138: 1:0.0720790:...,...