Visualise RNA-seq data in a the form of a sashimi plot

plot_sashimi plots the splicing events and coverage over a specific genes/transcript and/or genomic region of interest. The plotting are built from ggplot2 functions.

plot_sashimi(
  junctions,
  ref,
  gene_tx_id,
  case_id = NULL,
  control_agg_func = mean,
  region = NULL,
  annot_colour = c(ggpubr::get_palette("jco", 1), ggpubr::get_palette("npg", 7)[c(1, 3,
    2, 5, 6)], ggpubr::get_palette("jco", 6)[c(3)]),
  digits = 2,
  count_label = TRUE
)

Arguments

junctions
junction data as a RangedSummarizedExperiment-class object.

ref
either path to gtf/gff3 or object of class TxDb-class.

gene_tx_id
gene name as ensembl id, ensembl transcript id or gene name.

case_id
list of one element. This must be a character vector containing the

control_agg_func
list of one element. This must be a character vector containing the

region
a GenomicRanges of length 1 that is used to filter the exons/junctions. Only those that overlap this region are plotted.

annot_colour
character vector of colours for junction types. One value must be supplied labelling each of the 7 possible types.

digits
used in round, specifying the number of digits to round the junction counts to for visualisation purposes.

count_label
logical value specifying whether to add label the count of each junction.

Value

ggplot displaying the splicing (and coverage) surrounding the transcript/region of interest.

Examples

if (!exists("ref")) {
    # use Genomic state to load txdb (GENCODE v31)
    ref <- GenomicState::GenomicStateHub(version = "31", genome = "hg38", filetype = "TxDb")[[1]]
    # convert seqlevels to match junctions
    seqlevels(ref) <- stringr::str_replace(seqlevels(ref), "chr", "")
}

#> snapshotDate(): 2020-09-03

#> loading from cache

if (!exists("junctions_processed")) {
    junctions_processed <-
        junction_process(
            junctions_example,
            ref,
            count_thresh = c("raw" = 5),
            n_samp = c("raw" = 1),
            width_range = c(25, 1000000),
            types = c("ambig_gene", "unannotated"),
        )
    GenomeInfoDb::seqlevels(junctions_processed) <-
        paste0("chr", GenomeInfoDb::seqlevels(junctions_processed))
}

#> [1] "# Filtering junctions -----------------------------------------------------"
#> [1] "2020-10-14 17:58:36 - Filtering junctions..."
#> [1] "2020-10-14 17:58:36 - by count..."
#> [1] "2020-10-14 17:58:36 - by width..."
#> [1] "2020-10-14 17:58:36 - done!"
#> [1] "# Annotating junctions ----------------------------------------------------"
#> [1] "2020-10-14 17:58:36 - Obtaining co-ordinates of annotated exons and junctions from gtf/gff3..."
#> [1] "2020-10-14 17:58:52 - Getting junction annotation using overlapping exons..."
#> [1] "2020-10-14 17:58:53 - Tidying junction annotation..."
#> [1] "2020-10-14 17:58:53 - Deriving junction categories..."
#> [1] "2020-10-14 17:58:54 - done!"
#> [1] "# Filtering junctions -----------------------------------------------------"
#> [1] "2020-10-14 17:58:54 - Filtering junctions..."
#> [1] "2020-10-14 17:58:54 - by type..."
#> [1] "2020-10-14 17:58:54 - done!"
#> [1] "# Normalise junctions -----------------------------------------------------"
#> [1] "2020-10-14 17:58:54 - Clustering junctions..."
#> [1] "2020-10-14 17:58:54 - Normalising junction counts..."
#> [1] "2020-10-14 17:58:54 - done!"
#> [1] "# Score junctions ---------------------------------------------------------"
#> [1] "2020-10-14 17:58:54 - Calculating the direction of change of junctions..."
#> [1] "2020-10-14 17:58:54 - Generating junction abnormality score..."
#> [1] "2020-10-14 17:58:54 - done!"

sashimi_plot <- plot_sashimi(
    junctions = junctions_processed,
    ref = ref,
    gene_tx_id = "ENSG00000241973.10"
)

#> Warning: Removed 108 row(s) containing missing values (geom_path).

#> Warning: Removed 4 rows containing missing values (geom_label_repel).

#> Warning: Removed 108 row(s) containing missing values (geom_path).

#> Warning: Removed 4 rows containing missing values (geom_label_repel).